// frontend/lib/api/generated/model/recetaIngredienteIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface RecetaIngredienteIn {
  alimento_id: number;
  /** @exclusiveMinimum 0 */
  cantidad_bruta_g: number;
  /** @exclusiveMinimum 0 */
  cantidad_neta_g: number;
  unidad_id: number;
  /** @exclusiveMinimum 0 */
  factor_conversion?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  merma_pct?: number;
}